requirements.txt
cryptography>=42.0.0
argon2-cffi>=23.1.0
Pillow>=10.0.0
